fix(briefing): stop caching generic-fallback mush + readable date format
Root cause of "ari briefing is generic mush": when Groq's JSON output
fails to parse (or parses but with empty fields), the briefing endpoint
was writing canned filler to manifest.ariIntelligence.lastBriefing:
  "Lead data insufficient for situation assessment"
  "Contact needed to gauge motivation level"
  "Schedule call to assess opportunity and build rapport"
  "Motivation assessment unavailable"
  "Strategy unavailable"

Once cached, the UI could not distinguish this from real output. The
briefing stayed "mush" across reloads until briefingStale flipped —
and the stale flag only flips on new transcripts, which the pipeline
wasn't adding (fixed in earlier commits).

Now: parse/extraction failure throws. The endpoint returns 500 and the
UI can show a retry/error state. Garbage never gets cached.

Also: Next Action date label upgraded from cramped "Wed, Apr 29" to the
full "Wednesday, April 29" format (only for dates >7 days out; within
a week still reads as "Tomorrow" / "Monday" / etc.).

+++ src/components/leads/next-action.tsx
@@ -73,7 +73,8 @@ function formatDateLabel(iso: string | null): string | null {
   if (diffDays === -1) return 'Yesterday'
   if (diffDays > 0 && diffDays <= 6) return d.toLocaleDateString('en-US', { weekday: 'long' })
   if (diffDays < 0 && diffDays >= -6) return `${Math.abs(diffDays)}d ago`
-  return d.toLocaleDateString('en-US', { weekday: 'short', month: 'short', day: 'numeric' })
+  // Full, readable format: "Wednesday, April 29" — no cramped abbreviations
+  return d.toLocaleDateString('en-US', { weekday: 'long', month: 'long', day: 'numeric' })
 }
 
 function formatTimeLabel(iso: string | null, dateOnlyHint?: boolean): string | null {
